[O5fNSvsN] Цены -- добавлен экспорт UiPriceCardProps в shared/ui -- все статичные виджета PriceList данные были пересены в price/ui/list/consts/.. -- добавлен ключ для перебора UiPriceCard
