+++ src/pages/main/ui/page/index.vue
@@ -1,10 +1,11 @@
 <template>
   <LayoutHeaderWidget />
   <MainPageSectionsWelcome class="mt-[10.65rem] h-[52.55rem]" />
+  <PriceListWidget />
 </template>
 
 <script setup>
 useHead({
-  title: 'X-Install'
-})
+  title: "X-Install",
+});
 </script>
