+++ src/widgets/price/ui/list/consts/android.services.ts
@@ -1,0 +1,51 @@
+export const androidServices: UiPriceCardProps[] = [
+  {
+    service: `Мотивированные\nустановки в Play Market`,
+    serviceAccent: `Мотивированные\nустановки`,
+    serviceDescription: `Выведение вашего приложения в топ выдачи по ключевым словам. Стоимость указана\nза одну мотивированную установку по одному\nключевому слову`,
+    guarantees: [
+      "Любое количество установок",
+      "Быстрое продвижение",
+      "Низкая цена",
+      "Персональное консультирование",
+    ],
+    price: "0.15",
+    isUnitPrice: true,
+  },
+  {
+    service: `Разработка\nприложений на Android`,
+    serviceAccent: `приложений на Android`,
+    serviceDescription: `Продажа готовых приложений для арбитража\nтраффика или разработка под заказ`,
+    guarantees: [
+      "Под любые источники трафика",
+      "Персональное консультирование",
+      "Быстрое получение результата",
+    ],
+    price: "2000",
+    isUnitPrice: false,
+  },
+  {
+    service: `Продвижение\nInApp`,
+    serviceAccent: `InApp`,
+    serviceDescription: `Разработка стратегии для продвижения\nс помощью рекламных площадок Unity,\nIntegral и другие`,
+    guarantees: [
+      "Уникальная стратегия под вашу нишу",
+      "Пополнение без скрытых %",
+      "Запуск рекламных компаний",
+    ],
+    price: "5000",
+    isUnitPrice: false,
+  },
+  {
+    service: `Продвижение\nGoogle (UAC)`,
+    serviceAccent: `Google (UAC)`,
+    serviceDescription: `Разработка стратегии\nпродвижения с помощью\nрекламных площадок Google `,
+    guarantees: [
+      "Разработка шаблонов",
+      "Креативов , Текстов",
+      "Запуск рекламных компаний",
+    ],
+    price: "10,000",
+    isUnitPrice: false,
+  },
+];
