+++ src/widgets/price/ui/list/consts/index.ts
@@ -1,0 +1,3 @@
+export {androidServices} from './android.services'
+export {iosServices} from './ios.services'
+export {tabs} from './tabs'
