+++ src/widgets/price/ui/list/consts/ios.services.ts
@@ -1,0 +1,67 @@
+export const iosServices: UiPriceCardProps[] = [
+  {
+    service: `Мотивированные\nустановки в AppStore`,
+    serviceAccent: `Мотивированные\nустановки`,
+    serviceDescription: `Выведение вашего приложения в топ выдачи по ключевым словам. Стоимость указана\nза одну мотивированную установку по одному\nключевому слову`,
+    guarantees: [
+      "Любое количество установок",
+      "Быстрое продвижение",
+      "Низкая цена",
+      "Персональное консультирование",
+    ],
+    price: "0.24",
+    isUnitPrice: true,
+  },
+  {
+    service: `Разработка\nприложений на IOS`,
+    serviceAccent: `приложений`,
+    serviceDescription: `Продажа готовых приложений для арбитража\nтраффика или разработка под заказ`,
+    guarantees: ["Индивидуальный подход"],
+    price: "2999",
+    isUnitPrice: false,
+  },
+  {
+    service: `Продвижение\nInApp`,
+    serviceAccent: `InApp`,
+    serviceDescription: `Разработка стратегии для продвижения\nс помощью рекламных площадок Unity,\nIntegral и другие`,
+    guarantees: [
+      "Уникальная стратегия под вашу нишу",
+      "Пополнение без скрытых %",
+      "Запуск рекламных компаний",
+    ],
+    price: "5000",
+    isUnitPrice: false,
+  },
+  {
+    service: `Продвижение\nGoogle (UAC)`,
+    serviceAccent: `Google (UAC)`,
+    serviceDescription: `Разработка стратегии\nпродвижения с помощью\nрекламных площадок Google `,
+    guarantees: [
+      "Разработка шаблонов",
+      "Креативов , Текстов",
+      "Запуск рекламных компаний",
+    ],
+    price: "10,000",
+    isUnitPrice: false,
+  },
+  {
+    service: `Гарантированное\nпродвижение\nв топ по ключу`,
+    serviceAccent: `Гарантированное\nпродвижение`,
+    serviceDescription: `Стоимость формируется по результатам\nоценки каждого отдельного кейса`,
+    guarantees: [
+      "Оплата за результат",
+      "Быстрое продвижение",
+      "Уникальная система продвижения",
+    ],
+  },
+  {
+    service: `Бизнес\nсотрудничество`,
+    serviceAccent: `Бизнес`,
+    serviceDescription: `Мы открыты к взаимовыгодному сотрудничеству.\nРеализуем продвижение приложений в топы\nпо ключевым словам за долю в вашем приложении`,
+    guarantees: [
+      "Индивидуальный подход",
+      "Мы инвестируем в ваше развитие",
+      "Мы заинтересованы в успешном результате",
+    ],
+  },
+];
