+++ src/widgets/price/ui/list/consts/tabs.ts
@@ -1,0 +1,8 @@
+export const tabs = [
+  {
+    label: "IOS",
+  },
+  {
+    label: "Android",
+  },
+];
